aurabc-lb-validation/pyproject.toml
[project]
name = "aurabc-lb-validation"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "neo4j",
    "python-dotenv",
    "requests",
]
